Move the creat_app function to zhihudaily/app.py
